move service to service directory
