+++ central/main.go
@@ -140,9 +140,9 @@ import (
 	"github.com/stackrox/rox/central/tlsconfig"
 	"github.com/stackrox/rox/central/trace"
 	"github.com/stackrox/rox/central/ui"
-	usageService "github.com/stackrox/rox/central/usage"
 	usageCSV "github.com/stackrox/rox/central/usage/csv"
 	usageDataStore "github.com/stackrox/rox/central/usage/datastore"
+	usageService "github.com/stackrox/rox/central/usage/service"
 	userService "github.com/stackrox/rox/central/user/service"
 	"github.com/stackrox/rox/central/version"
 	vStore "github.com/stackrox/rox/central/version/store"
